Add lexer support for delimiters
The lexer is now capable of using whitespace as a delimiter to separate identifiers. For example, before adding this support, the lexer would treat "5555xxxx" as two separate legal tokens, {Int 5555} and {Ident xxxx}. This however is wrong as the lexer should treat this as one large illegal token: {Illegal 5555xxxx}. The intended behavior is to treat whitespace as a delimiter within the s expression.