Add expression parsing
The parser is now able to parse Atoms, and some list expressions
like binary expressions (like adding two integers). The parser still
lacks support for most planned language features, but this serves as a
start. In addition, the Token struct was also given it's missing
Line and Col fields which were mistakenly omitted in the last commit.