Add parsing for cond statements
In addition, several changes were made to parser.go to enhance the
readability of the file.

- All uses of the expectPeek function were changed to expectCur.
Before this change, it wasn't clear why expectPeek was to be used
over expectCur, and readability was damaged without clear uses cases
for both. Now, expectPeek should only be used when a peek is strictly
necessary.
- All specific parsing expressions now ensure that the current token
aligns with what the function expects. In addition, all parsing functions
now advance to the next token when complete.
- All manual checks for delimiters are now replaced with the eatDelimiter()
funtion which ensures a delimiter is present, and advances past it.

By improving readability with these changes, it should be easier to
reason what the parser is doing at each step in the parsing process.