Remove literals from delimiter tokens
In the Simple programming language, any kind of whitespace of any
length is a valid delimiter. The parser only checks that a delimiter
is present, but does nothing with its literal. By placing an empty
string for every delimiter's literal (effectively removing it) it
becomes easier for test functions to ensure got delimiters are equal
to expected.